Update sign-in-teams.md
We are improving the sign in experience to Microsoft Teams. Here is a quick overview of the changes:
• Users that have a cloud username that is different from their on-prem username can now sign into Teams without the username pre-filled on Welcome screen

IT admins need to set up a registry key that will ignore the on-prem username and not prefill it on the Welcome screen at the time of sign in. However, if the on-prem username ends in .local or .corp, admins do not need to take any explicit action to suppress the prefill.